requirements.txt
numpy==1.23.3
pandas==1.5.0
Pillow==9.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
streamlit==1.12.2
plotly==5.3.1
